Use maven-war-plugin to reduce package size
This reduces the resulting package size from about 8.6MB to 1.7MB.
If some files are missing or should be included in package, add them
into <warSourcesIncludes> section in pom.xml.

Attention! With this change the resultung package will be prepared for
production use. No 'developer.html' or other developer files will be
present in package.