Change displayed value and sorting of data store depending on screen with
This should display a courses shortname ond order store by shortname if
the displays width is smaller or equal than 321 pixel. If the display has
more than 321px, do the same with the courses full name.

The value of 321px is used at different other locations in code, so I
decided to use this value too.