Fixed #5955: localStorage can handle strings only
The server does not handle empty strings for courseId and courseType
correctly, so we needed to send NULLs. But localStorage is for
strings only, and NULL gets translated into the string "null."

This causes the implementation of the course connector to think
that "null" is the type and id of a valid course session. Hence,
questions get assigned the "releasedFor": "course" attribute.