Enabled preparation questions
Open Issue: Currently, depending on which button the user has
clicked, the "mode" of the QuestionPanel changes. The consequence
is that if the user clicks on the question button on the tab bar,
she will see the question mode she last selected (i.e. the
preparation questions are shown instead of the lecture questions
if the user had selected lecture questions on the InClass panel).