Start server with Grunt run task
A web server for static resources and a proxy to the backend API are
now started with the run task.

ARSnova Mobile no longer relies on the backend to provide the frontend
application files. In consequence, the development environment is more
flexible. The local repositories can be located independently and it is
possible to deploy the backend as WAR file on an application server
instead of using the Maven jetty:run goal. This should ease the setup
process for pure frontend developers.

Additionally, the newly provided development environment is much closer
to the recommended production setup. The frontend is located at /mobile
while the API is accessible at /api.