Use the default session when input textfield "joinSessionInput" is empty.
This has been implemented due to a user review. The users has mentioned
that it would be easier to call the default session when no textinpunt is
given.

At the moment this implementation can be best seen as a (dirty) hack, because of a
bug or some changes wich are not finished yet. The problem is that the
global property "ARSnova.app.globalConfig.demoSessionKey" returns undefined.

According to that, the Label which shows the user the id for the defaul session
doenst show up at the moment. But that is not in the scope of this pull request.