Clarify commands for continuous build
The aim of this change is to better communicate
the steps needed to run ARSnova mobile and backend
continuously. The essential steps are to first
start the Jetty server, and then to use Grunt to
start a proxy and to watch the file system.

Previously, the section was a bit ambiguous as
to what the exact order of the commands is.
For example, `grunt build` is not part of the continuous
build process, but was mentioned in this section.