Task #15810: Added two action buttons to showcaseEditButtons.
Due to width limitations on mobile devices, the two additional buttons
will be shown in an extra row - below the existing buttons. On
devices/browser windows with a screen width >= 620px all five buttons
will be shown in one row.

The rendering of the two rows/one row-mode is dynamic and will update
automatically on window resize.