Task #15893: Add scaling feature for showcase question carousel.
This feature commit introduces a text scaling option for speakers
showcase question carousel. When the browser window width is >=700px, a
iconed button will be rendered on the bottom right of the carousel. This
buttons reveals a slider field by which the font-size of question text
and answer options can be scaled.